consumer/requirements.txt
flask
kafka-python
requests
apscheduler
pymongo
python-dotenv
